Mods to make "interactive input" work with awk
User @davidar (github.com/davidar, David A Roberts) notes that wak/toybox awk
did not work correctly with interactive input. His example was

wak '/^say/ {print $2}

Taking input from the terminal, input of 'say hi' should respond 'hi', but it
was not working correctly due to the behavior of the input routines not
stopping at a newline, but attempting to fill a large buffer to accommodate the
use of a regex as RS (record separator).

David provided a patch using fgets() to stop reading input at a newline, but it
caused the RS to be ignored. I tried checking for RS == "\n" but that was not
sufficient, as RS is usually a newline, and causing all input to go to fgets()
was a big performance hit. So I also added a field 'is_tty' for each opened
file (including stdin), and only use fgets() if RS is "\n" and also if the
input file isatty(). Then I find that I also have to set toybox awk to use line
buffering. When using (non-toybox) wak, the system sets the output buffering to
non-full-buffering if stdout is an "interactive device" (as required by POSIX).

It would be nice if toybox provided an option to not set the buffering for
stdout at all, say TOYFLAG_DEFAULTBUF, leaving it to the system, then toybox
awk would not take the performance hit on stdout for all file operations.