Mod getrec_f() to permit null RS="()"
An empty RS string (RS="") is special and causes awk to enter multiline record
mode. But a null RS (e.g. RS="()") matches an empty string at the beginning of
the file, and cannot match any separator string, and caused an infinite loop of
output. Other awks (except busybox awk, which has the same bug, and mawk which
gives an error on that regex) treat it as not matching anything (same as
RS="^$), so entire input becomes a single record (../cmit_ggdb). Here we adopt the same
policy.
Also added CHANGELOG.md and updated README.md.